api: http: do replace 404 with 204 on DELETE if device was not found.
Do not return 404 if DELETE on /devices/{id} did not find any resources to
delete. In the end, the effect of DELETE if there was no device and if one was
removed is exactly the same (device is no longer present), but non-success
status may confuse some integration tools.

Changelog: none
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>