tests: create fewer groups in acceptance tests
When creating a large number of groups within a short period of time will cause
heavy logging. This may in turn cause docker shim to get stuck pushing the logs
from the container outside. A workaround is to either manually `cat` on the
stdout file descriptor of docker shim or reduce logging.

Changelog: None

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>