store/mongo: refactor failing tests
tests relying on exact mongo error messages started failing with a patch
update of the mongo 3.4 package in trusty/multiverse; the messages are
different now, and in general shouldn't be relied on.

refactor to avoid always comparing against an exact error - IsError signals
that an error occured, and OutputError will be set optionally, if
we mean to do an exact comparison.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>