store: migrate to globalsign/mgo
this is a hail mary attempt to fix unit tests failing on master.
the issue was spotted while refactoring the travis pipeline. despite a
lot of debugging, the cause was not found.

the update didn't help, but i'm leaving this here because we'd do
it anyway at some point.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>