vendor: update stretchr/testify to get assert.ElementsMatch
there is a heisenbug in some of our tests where we
rely on the order of 'expected' slices.

the order may very well be random - if somewhere along the way
maps are involved (even under the hood). if in such a case
we do assert.Equal - the tests may or may not pass.

updated lib contains 'ElementsMatch', which ignores the order, and
that's what we should use.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>