api/http/tests: use assert.ElementsMatch instead of Equals on 'expected' slices
in some cases, the order of output slices is not guaranteed - if
processing involves map iteration on any level (our/stdlib/3rd party code).

url parsing is one of those - query string params are read into a
map before we extract filters out of them.

assert.Equals is sensitive to order, and useless in this case (it
*can* pass for some percentage of the runs though).

assert.ElementsMatch is the order-insensitive replacement.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>