store/tests: fix heisenbug in TestUpsertAttributes
there is a definite bug - in 2 cases, expected 'slice'
attributes are defined as []string.

the received values are *always* []interface{}, so this will never
DeepEqual in the compare() func. the string slice has to be converted to
interface slice ('Values of distinct types are never deeply equal.' -
spec).

the reason it pops randomly is because of *another* bug:
in the compare func, we return 'true' prematurely when descriptions are nil.
we should never exit early with success because we're iterating through all of
device's attributes - it's plain and simple a bug. in the loop, we can
only exit early if a comparison is unsuccessful.

since the iteration is over a map - sometimes the early exit happens,
sometimes it doesn't - so we either hit the first bug or we don't.

all is fixed now, and btw the description comparison is converted to
DeepEqual (this is a good case).

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>