tests: adjust to v1 bugfixes
for searching and sorting - we don't need the scope prefixes.
the idea was to have v1 operate transparently on 'inventory' scope.
this means:
- v1/PATCH accepts scopes - inserts as 'inventory-<name>'
- GET can filter on just 'name' - and prepends 'inventory-' internally
- same for sort param

after the bugfix, where the new helpers are actually used, filtering and
sorting tests must be adjusted - scope prefix is out of the filter/sort
param.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>