api/http: GET management/v2/devices
uses the 'unified' url, 1:1 with gateway urls.
a change in nginx.conf will follow.

re: tests:
i don't want to fiddle with mockListDevices and mockListDevicesV1
anymore.

we can easily create predefined devs with attribute combinations
specific for a given test and be done - without thinking
if we're not breaking something that's used in N other tests, for
different api versions.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>