store/mongo: Create 'group', 'updated_ts' and 'created_ts' attributes
These attributes are reflecting the corresponding properties in the root
of the structure.
We need to keep both document elements for now until next micro-release
as the state before and after migrating the db needs to co-exist for a
short time. The places that needs modification to complete the migration
is marged with 'TODO'.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>