Refactor datastore interface combining single and bulk operations
This commit removes many of the .*One mongo operations from the
datastore, substituting it with a single .*Many interface, letting the
app layer (inv package) perform the abstraction.
This was motivated by the modified api handler looping over all items in
an array, repeating calls to the storage layer. This handler was also
changed to use a bulk interface.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>