Do not update the updated_ts timestamp in internal end-points
The deviceauth service consumes the internal management API end-points
to update the inventory fields. In this case, we don't want to update
the update_ts timestamp. We provide a new method in the DataStore
(UpsertDevicesAttributesWithUpdated) which explicitly updates the
update_ts field, and we use this method from the device API end-points
consumed by the Mender Client to update the device inventory data.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>